mds: Don't blocklist the clients in any replay state
The mds in 'standby-replay' state should not blocklist clients.
Attempting to do so would hit an assert while journaling.

Before the commit 79f7576 the 'Server:apply_blocklist' used to
fetch the blocklist addresses from 'Objecter::consume_blocklist_events'
which returns blocklist addresses only if 'blocklist_events_enabled'
are enabled. This is not enabled for mds in 'standby-replay' state.
Hence this issue wasn't seen.

After the commit 79f7576 the 'Server::apply_blocklist' fetches the
blocklist addresses from the osdmap. This would return the blocklist
addresses irrespective of the mds state. The mds in 'replay' state
doesn't blocklist the clients but if the mds is in 'standby-replay'
state, it tries to journal the blocklist and asserts. This patch
fixes the same by not allowing the blocklist if the mds is in any
replay state.

Introduced-by: 79f7576401cc9d857f84396314d7476336c0e271
Fixes: https://tracker.ceph.com/issues/56012
Signed-off-by: Kotresh HR <khiremat@redhat.com>
(cherry picked from commit 733fb72acfeb79adbb26166ca13b1e5927cd6547)