rgw: implement initial flight server functionality
Implements the ability for a flight to be created when the object is
retrieved by an S3 get.

Adds FlightServer abilities ListFlights, GetFlightInfo, GetSchema, and
DoGet. Adds an interface for a store for flight information and adds
an in-memory implemtation of it.

This code is functionality is early-stage and lacks some planned
efficiencies.

Signed-off-by: J. Eric Ivancich <ivancich@redhat.com>