rgw_sal_motr: [CORTX-32423] fix listing of same-prefix objs (#326)
Currently, the format of null-version objects keys stored in
bucket index in Motr does not have any suffix. But versioned
objects do have the suffix: "key[instance]". This leads to the
situations when null-objects can be fetched separately from
their versioned objects, in case there are some other objects
with the same key prefix. For example, consider the following
case:

    obj1
    obj1-2
    obj1[instance]

It's because "-" symbol goes before "[" in the ASCII-table.
As a result, the output of list-object-versions command for
such cases will be broken.

This also may affect the correct setting of IsLatest flag for
the object versions in the versioned bucket, when the new
version is put into the bucket or some existing one is deleted
from the bucket. It's because when we find the latest object
version to update, we fetch two versions using the NEXT Motr
index operation starting from the object name key (see
fetch_latest_obj()). So in the above example, obj1 will always
be returned as the latest version, which may be wrong.

Solution: append '\a' symbol to the null-version object keys,
and use it to separate key name from the instance hash.
'\a' non-printable symbol in the ASCII-table goes before all
the allowed printable symbols in the key name. So the above
example will be fetched from Motr KV like this:

    obj1\a
    obj1\ainstance
    obj1-2\a

Kudos to Sining Wu <sining.wu@seagate.com> for reporting this.

Closes #324.

Signed-off-by: Andriy Tkachuk <andriy.tkachuk@seagate.com>