client: switch to use 32 bits ext_num_fwd
The MDS will increase the forward count, and if the forward count
is less than the one saved in request in client side, that means
the MDS is old version and it was overflowed. Then just stop
forwarding.

Fixes: https://tracker.ceph.com/issues/57854
Signed-off-by: Xiubo Li <xiubli@redhat.com>
(cherry picked from commit 3584ecc876b981b2204fbd7f720867e454639c58)
Conflicts:
- src/client/Client.cc: there is no -CEPHFS_EMULTIHOP