crimson/osd: fix ignoring --no-mon-config
Before we were treating the `--no-mon-config` as one of the
`seastar_n_early_args`. However, this was wrong as it truly
belongs to `config_proxy_args` as:

```cpp
int md_config_t::parse_argv(ConfigValues& values,
                            const ConfigTracker& tracker,
                            std::vector<const char*>& args, int level)
{
    // ...
    else if (ceph_argparse_flag(args, i, "--no-mon-config", (char*)NULL)) {
      values.no_mon_config = true;
    }
    // ...
}
```

The net result of this ignoring `--no-mon-config` which was
the reason behind many dead jobs at Sepia.

Signed-off-by: Radoslaw Zarzynski <rzarzyns@redhat.com>