mgr/telemetry: reset health warning after re-opting-in
When telemetry requires re-opting-in (either whenever new collections
which require nagging are available, or on major upgrades) a health
warning is set by the module. This health warning should be reset once
the user re-opts-in (with `ceph telemetry on`), but currently it might
take longer. Fixing it here by waking up serve() immediately after
re-opting-in, which will invoke refreshing health checks.

Fixes: https://tracker.ceph.com/issues/56486
Signed-off-by: Yaarit Hatuka <yaarit@redhat.com>