mgr/cephadm: recreate osd config when redeploy/reconfiguring
OSDs have a config file that includes addresses for the mon daemons.
We already have in place logic to cause a reconfig of OSDs if the mon map
changes, but when we do we aren't actually regenerating the config
so it's never updated with the new mon addresses. This change is to
have us recreate the OSD config when we redeploy or reconfig an OSD
so it gets the new mon addresses.

Fixes: https://tracker.ceph.com/issues/57007

Signed-off-by: Adam King <adking@redhat.com>
(cherry picked from commit 13358f52e0e1b57fa0e7afc27956e1503a373fc2)

Conflicts:
	src/pybind/mgr/cephadm/module.py