rgw: improved admin token diagnostics.
As a result of assorted bugs, various versions of radosgw have been caught
trying to use a defective cached admin token.  The resulting behavior
is wrong and the log messages are not very clueful.  This fix includes
logic to prevent this; if a token is bad, it should not be cached,
and should generate a more obvious error message.

This fix:
/1/ be way pickier about incoming admin tokens.
/2/ try to produce more useful error messaging upon misbehavior or misconfiguration.

Fixes: https://tracker.ceph.com/issues/61258

Signed-off-by: Marcus Watts <mwatts@redhat.com>