qa/import-legacy: install python3 package for nautilus ceph
The 'volume_client' script will use the python3 and will import
the'ceph_volume_client' volume, we need to install the python3
packages when installing the nautilus ceph.

Or it will fail to generate the keyring contents for the
'client.vol_data_isolated', and then when authenticating from mon
the mon will only receive a corrupt payload.

Fixes: https://tracker.ceph.com/issues/57083
Signed-off-by: Xiubo Li <xiubli@redhat.com>