rgw: fix garbage in log line upon authentication failure w/ rgw_s3_auth_use_keystone
When s3 authentication against keystone fails, a log message is produced
that ends with a bl.c_str() string.  Typewise this is a char *, but it's not
a well-formatted C string because there is no guarantee it ends in a nul.
In practice, it frequently contains unrelated data from some previous
use of the memory.  This commit converts it into a std::string
using bl.length() to set the length right.

Fixes: https://tracker.ceph.com/issues/57093

Signed-off-by: Marcus Watts <mwatts@redhat.com>