doc/rados: add ec-pool creation procedure
This PR filters Casey Bodley's expertise through my ignorance in order
to produce a two-step tutorial on how to create an erasure-coded pool
and how to associate that pool with the "rgw" application.

This is the first in what will probably be a long series of edits to
the RADOS documentation.

Signed-off-by: Zac Dover <zac.dover@gmail.com>