mgr/cephadm: support for max_mds setting in mds specs
Users are able to set the max_mds setting for an fs
in order to set how many mds daemons should be active.
So "ceph fs set <fs-name> max_mds 2" would tell the fs
to have 2 mds daemons be active and the rest standby.
This change allows setting this though the mds service
spec and also the "ceph fs volume create" command.

Signed-off-by: Adam King <adking@redhat.com>