docs: correct zone rm command in docs to zone delete
Signed-off-by: Salar Nosrati-Ershad <snosratiershad@gmail.com>

fix: correct zone rm command in docs to zone delete

doc/rados/operations: add prompts to operating.rst

This commit adds ".. prompt:: bash $"-style prompts to operating.rst.
This brings this file up to the standard established in 2020 when
Kefu added support for the ".. prompt::" directive.

This commit is a part of an initiative to modernize the presentation
of all BASH commands in the RADOS documentation.

The progress of this project can be tracked here:
https://tracker.ceph.com/issues/57108

Signed-off-by: Zac Dover <zac.dover@gmail.com>

seastore: prevent re-entering do_gc_cycle() when run_until_halt is called

Signed-off-by: Myoungwon Oh <myoungwon.oh@samsung.com>
Signed-off-by: Yingxin Cheng <yingxin.cheng@intel.com>

test/seastore: add write seqeunce to avoid racing

In the middle of test_random_write_concurrrent(), there is a chance
that consume(laddr A), regarding transaction 1, in try_submit_transaction
can be called after transaction 2's consume(laddr A) is invoked (note that
seastore handles both transaction 1 and 2 in order, and each other's laddr
are the same).

To avoid this, this commit adds a write seqeunce to prevent overwriting
old checksum associated with laddr---old checksum will not be updated
after write sequence comparison.

Signed-off-by: Myoungwon Oh <myoungwon.oh@samsung.com>

crimson/osd: Port rgw watcher list

Implemented function cls_cxx_list_watchers for crimson

Signed-off-by: Nitzan Mordechai <nmordec@redhat.com>