mgr/dashboard: set path for jsonnetfmt
in 4a3afcf27769a512fba7e0194d2e3b6a55e06bca, the $PATH is set for
the test, but we cannot set multiple properties with a single
`set_property()` cmake comamnd. also, we should not override the
all `ENVIRONMENT` variables with a single `PATH` env variable.

in this change, because to manipulate the "PATH" in `ENVIRONMENT`
property of the test target would be complicated in comparison with
to add yet another enn variable, a new env variable is used for
the directory in which `jsonnetfmt` is located.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>